Add Redis and Prometheus support

+++ modules/config/dsl/src/main/kotlin/RedisConfig.kt
@@ -24,18 +24,18 @@ data class RedisConfig(
     val sentinels: List<String> = listOf(),
     val master: String? = null,
     val password: String? = null,
-    val index: Long = 5,
+    val index: Int = 5,
     val host: String = "localhost",
-    val port: Long = 6379,
+    val port: Int = 6379,
     val ssl: Boolean = false
 ) {
     class Builder: org.noelware.charted.common.Builder<RedisConfig> {
         private var sentinels = mutableListOf<String>()
         var masterName: String? = null
         var password: String? = null
-        var index: Long = 5
+        var index: Int = 5
         var host: String = "localhost"
-        var port: Long = 6379
+        var port: Int = 6379
         var ssl: Boolean = false
 
         fun addSentinel(host: String, port: Int): Builder = addSentinel("$host:$port")
